requirements.txt
pytest==9.0.2
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.10.0+cpu
torchaudio==2.10.0+cpu
speechbrain==1.0.3
webrtcvad-wheels==2.0.14
onnx==1.20.1
onnxruntime==1.24.4
